Producer/requirements.txt
python-dotenv
websocket-client
finnhub-python
pyzmq
orjson